+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 # Changelog #
 
 
-## 1.8.0 ??? (unreleased)
+## 1.8.0 Saracenia Purpurea (unreleased)
 
 ### Features
 - Improve timeline resource.
 - Add sitemap of taiga-front (the web client).
 - Search by reference (thanks to [@artlepool](https://github.com/artlepool))
 - Add call 'by_username' to the API resource User
+- i18n.
+  - Add deutsch (de) translation.
+  - Add nederlands (nl) translation.
 
 ### Misc
 - Lots of small and not so small bugfixes.